codigo/Live265/take_1/pyproject.toml
[tool.poetry]
name = "take-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-distro = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
